refactor: 예외 발생 시 크롤링하던 로직에서 빈 문자열 저장하도록 변경
