+++ backend/baguni-api/src/main/java/baguni/api/service/user/service/strategy/RankingInitStrategy.java
@@ -11,8 +11,8 @@ import org.springframework.stereotype.Component;
 import baguni.api.service.user.dto.UserInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import baguni.api.service.link.dto.LinkInfo;
-import baguni.api.service.link.exception.ApiLinkException;
+import baguni.entity.model.link.dto.LinkInfo;
+import baguni.entity.exception.link.ApiLinkException;
 import baguni.api.service.link.service.LinkService;
 import baguni.api.service.pick.dto.PickCommand;
 import baguni.api.service.pick.service.PickService;
@@ -64,7 +64,7 @@ public class RankingInitStrategy implements ContentInitStrategy {
 			try {
 				linkInfo = linkService.getLinkInfo(curr.url());
 			} catch (ApiLinkException exception) {
-				linkInfo = linkService.saveLinkAndUpdateOgTagByJsoup(curr.url());
+				linkInfo = linkService.saveLink(curr.url()); // url 외에 다른 필드는 모두 빈 문자열인 Link 생성
 			}
 			if (linkInfo.title().isBlank()) {
 				continue;
